requirements.txt
# HubSpot integration
hubspot-api-client
python-dotenv

# Google Sheets integration
gspread
google-auth

# Data manipulation
pandas

# Development
pytest
black
nox
